refactor: move SLA evaluation engine + trend endpoint into proprietary/inference-latency-pro
Addresses PR #124 review: Pro-gated parts of the inference-latency feature
were sitting in apps/ instead of following the proprietary module pattern
used by anomaly-detection / key-analytics / webhook-pro.

What moved out of apps/api/src/inference-latency/:
- sla.ts (the debounce engine) -> proprietary/inference-latency-pro/sla.ts
- evaluateSlas() + slaState + dispatchInferenceSlaBreach hook
  -> proprietary/inference-latency-pro/inference-latency-pro.service.ts
- GET /inference-latency/trend (Pro-gated via LicenseGuard)
  -> proprietary/inference-latency-pro/inference-latency-pro.controller.ts
- sla.spec.ts and sla-state-gc.spec.ts moved alongside

What stays in apps/ (and why):
- getProfile + computeProfile + getTrend computation: the math is open;
  only exposure of the trend endpoint is Pro-gated.
- inference_sla_breach Prometheus gauge registration: the gauge lives
  on the OSS PrometheusService, just like betterdb_anomaly_events_total
  and friends. The Pro service is the only caller of
  updateInferenceSlaBreachMetrics, matching how proprietary/anomaly-detection
  drives anomaly Prometheus metrics.
- inference_sla_config column on app_settings: schema stays in OSS, same
  as the anomaly_events / correlated_anomaly_groups tables and the
  anomaly_poll_interval_ms column. The Pro service is the only reader.

Hook contract:
- New token INFERENCE_LATENCY_PRO_SERVICE + IInferenceLatencyProService
  interface in @betterdb/shared.
- OSS InferenceLatencyService injects it as @Optional and calls
  onProfileTick(ctx, profile) after each poll tick. Without the Pro
  module loaded, no SLA evaluation runs and no breach gauge updates fire.
- onConnectionRemoved is forwarded so debounce state is GC'd when
  connections drop.

Conditional load wiring:
- app.module.ts adds an InferenceLatencyProModule try/require alongside
  WebhookProModule and registers it in proprietaryImports.

Tests: 43 pass in apps/api/src/inference-latency, 22 pass in
proprietary/inference-latency-pro (sla, sla-state-gc, controller).
