+++ apps/api/src/inference-latency/__tests__/inference-latency.controller.spec.ts
@@ -1,23 +1,13 @@
 /* eslint-disable @typescript-eslint/no-explicit-any */
 import { BadRequestException, NotFoundException } from '@nestjs/common';
 import { Test, TestingModule } from '@nestjs/testing';
-import { LicenseGuard } from '@proprietary/licenses';
 import { ConnectionRegistry } from '../../connections/connection-registry.service';
 import { InferenceLatencyController } from '../inference-latency.controller';
-import {
-  InferenceLatencyService,
-  InferenceLatencyValidationError,
-} from '../inference-latency.service';
-
-class AllowingGuard {
-  canActivate(): boolean {
-    return true;
-  }
-}
+import { InferenceLatencyService } from '../inference-latency.service';
 
 describe('InferenceLatencyController', () => {
   let controller: InferenceLatencyController;
-  let service: { getProfile: jest.Mock; getTrend: jest.Mock };
+  let service: { getProfile: jest.Mock };
 
   const buildModule = (registry: Partial<ConnectionRegistry>): Promise<TestingModule> =>
     Test.createTestingModule({
@@ -26,10 +16,7 @@ describe('InferenceLatencyController', () => {
         { provide: InferenceLatencyService, useValue: service },
         { provide: ConnectionRegistry, useValue: registry },
       ],
-    })
-      .overrideGuard(LicenseGuard)
-      .useClass(AllowingGuard)
-      .compile();
+    }).compile();
 
   const registry = (
     known: string[],
@@ -45,7 +32,6 @@ describe('InferenceLatencyController', () => {
   beforeEach(async () => {
     service = {
       getProfile: jest.fn().mockResolvedValue({ buckets: [] }),
-      getTrend: jest.fn().mockResolvedValue({ points: [] }),
     };
     const module = await buildModule(registry(['conn-1'], 'default-conn') as any);
     controller = module.get(InferenceLatencyController);
@@ -103,54 +89,6 @@ describe('InferenceLatencyController', () => {
       await expect(
         fallbackController.getProfile(undefined, undefined, undefined, undefined),
       ).rejects.toBeInstanceOf(NotFoundException);
-    });
-  });
-
-  describe('GET /trend', () => {
-    it('requires bucket, startTime, and endTime', async () => {
-      await expect(
-        controller.getTrend(undefined, '1', '2', undefined, 'conn-1'),
-      ).rejects.toBeInstanceOf(BadRequestException);
-      await expect(
-        controller.getTrend('FT.SEARCH:idx', undefined, '2', undefined, 'conn-1'),
-      ).rejects.toBeInstanceOf(BadRequestException);
-      await expect(
-        controller.getTrend('FT.SEARCH:idx', '1', undefined, undefined, 'conn-1'),
-      ).rejects.toBeInstanceOf(BadRequestException);
-    });
-
-    it('rejects non-numeric time params', async () => {
-      await expect(
-        controller.getTrend('FT.SEARCH:idx', 'not-a-number', '2', undefined, 'conn-1'),
-      ).rejects.toBeInstanceOf(BadRequestException);
-    });
-
-    it('rejects non-positive bucketMs', async () => {
-      await expect(
-        controller.getTrend('FT.SEARCH:idx', '1', '2', '-5', 'conn-1'),
-      ).rejects.toBeInstanceOf(BadRequestException);
-    });
-
-    it('delegates valid input to the service', async () => {
-      await controller.getTrend('FT.SEARCH:idx', '100', '900', '60000', 'conn-1');
-      expect(service.getTrend).toHaveBeenCalledWith('conn-1', 'FT.SEARCH:idx', 100, 900, 60000);
-    });
-
-    it('translates InferenceLatencyValidationError to 400', async () => {
-      service.getTrend.mockRejectedValueOnce(
-        new InferenceLatencyValidationError('bin cap exceeded'),
-      );
-      await expect(
-        controller.getTrend('FT.SEARCH:idx', '100', '900', undefined, 'conn-1'),
-      ).rejects.toBeInstanceOf(BadRequestException);
-    });
-
-    it('lets non-validation service errors propagate (5xx path)', async () => {
-      const boom = new Error('ECONNRESET');
-      service.getTrend.mockRejectedValueOnce(boom);
-      await expect(
-        controller.getTrend('FT.SEARCH:idx', '100', '900', undefined, 'conn-1'),
-      ).rejects.toBe(boom);
     });
   });
 });
