+++ apps/api/src/inference-latency/inference-latency.module.ts
@@ -1,13 +1,12 @@
 import { Module } from '@nestjs/common';
 import { ConnectionsModule } from '../connections/connections.module';
 import { PrometheusModule } from '../prometheus/prometheus.module';
-import { SettingsModule } from '../settings/settings.module';
 import { StorageModule } from '../storage/storage.module';
 import { InferenceLatencyController } from './inference-latency.controller';
 import { InferenceLatencyService } from './inference-latency.service';
 
 @Module({
-  imports: [StorageModule, ConnectionsModule, PrometheusModule, SettingsModule],
+  imports: [StorageModule, ConnectionsModule, PrometheusModule],
   controllers: [InferenceLatencyController],
   providers: [InferenceLatencyService],
   exports: [InferenceLatencyService],
