+++ proprietary/inference-latency-pro/__tests__/sla-state-gc.spec.ts
@@ -1,11 +1,9 @@
 /* eslint-disable @typescript-eslint/no-explicit-any */
 import { Test } from '@nestjs/testing';
-import { InferenceLatencyService } from '../inference-latency.service';
-import { ConnectionRegistry } from '../../connections/connection-registry.service';
-import { SettingsService } from '../../settings/settings.service';
-import { PrometheusService } from '../../prometheus/prometheus.service';
+import { PrometheusService } from '@app/prometheus/prometheus.service';
+import { SettingsService } from '@app/settings/settings.service';
 import type { InferenceLatencyBucket } from '@betterdb/shared';
-import type { ConnectionContext } from '../../common/services/multi-connection-poller';
+import { InferenceLatencyProService } from '../inference-latency-pro.service';
 
 function bucket(overrides: Partial<InferenceLatencyBucket> = {}): InferenceLatencyBucket {
   return {
@@ -20,21 +18,9 @@ function bucket(overrides: Partial<InferenceLatencyBucket> = {}): InferenceLaten
   };
 }
 
-describe('InferenceLatencyService SLA state GC', () => {
+describe('InferenceLatencyProService SLA state GC', () => {
   async function build(config: Record<string, { p99ThresholdUs: number; enabled: boolean }>) {
-    const storage = {
-      getCommandLogEntries: jest.fn().mockResolvedValue([]),
-      getSlowLogEntries: jest.fn().mockResolvedValue([]),
-      getVectorIndexSnapshots: jest.fn().mockResolvedValue([]),
-    };
-    const registry = {
-      get: jest.fn().mockReturnValue({
-        getCapabilities: () => ({ hasCommandLog: true }),
-        getConfigValue: jest.fn().mockResolvedValue('0'),
-      }),
-    };
     const prometheus = {
-      updateInferenceLatencyMetrics: jest.fn(),
       updateInferenceSlaBreachMetrics: jest.fn(),
     };
     const settings = {
@@ -42,32 +28,24 @@ describe('InferenceLatencyService SLA state GC', () => {
     };
     const module = await Test.createTestingModule({
       providers: [
-        InferenceLatencyService,
-        { provide: 'STORAGE_CLIENT', useValue: storage },
-        { provide: ConnectionRegistry, useValue: registry },
+        InferenceLatencyProService,
         { provide: PrometheusService, useValue: prometheus },
         { provide: SettingsService, useValue: settings },
       ],
     }).compile();
-    return { service: module.get(InferenceLatencyService), settings, prometheus };
+    return { service: module.get(InferenceLatencyProService), settings, prometheus };
   }
 
-  function ctx(connectionId = 'conn-1'): ConnectionContext {
-    return {
-      connectionId,
-      connectionName: 'local',
-      client: {} as any,
-      host: 'localhost',
-      port: 6379,
-    };
+  function ctx(connectionId = 'conn-1') {
+    return { connectionId, host: 'localhost', port: 6379 };
   }
 
-  function internalState(service: InferenceLatencyService): Map<string, unknown> {
+  function internalState(service: InferenceLatencyProService): Map<string, unknown> {
     return (service as any).slaState as Map<string, unknown>;
   }
 
   function runEvaluate(
-    service: InferenceLatencyService,
+    service: InferenceLatencyProService,
     connectionId: string,
     buckets: InferenceLatencyBucket[],
   ) {
@@ -80,7 +58,7 @@ describe('InferenceLatencyService SLA state GC', () => {
       buckets,
       generatedAt: Date.now(),
     };
-    return (service as any).evaluateSlas(ctx(connectionId), profile);
+    return service.onProfileTick(ctx(connectionId), profile);
   }
 
   it('clears state for an index that was configured then disabled', async () => {
@@ -88,11 +66,9 @@ describe('InferenceLatencyService SLA state GC', () => {
       idx_a: { p99ThresholdUs: 5_000, enabled: true },
     });
 
-    // First tick: breach, state entry created.
     await runEvaluate(service, 'conn-1', [bucket({ p99: 9_000 })]);
     expect(internalState(service).has('conn-1|idx_a')).toBe(true);
 
-    // User disables the SLA.
     settings.getCachedSettings.mockReturnValueOnce({
       inferenceSlaConfig: { idx_a: { p99ThresholdUs: 5_000, enabled: false } },
     });
@@ -140,13 +116,9 @@ describe('InferenceLatencyService SLA state GC', () => {
     const { service, prometheus } = await build({
       idx_a: { p99ThresholdUs: 5_000, enabled: true },
     });
-    // Tick 1: breach, state lands.
     await runEvaluate(service, 'conn-1', [bucket({ p99: 9_000 })]);
-    // Tick 2: no FT.SEARCH traffic.
     prometheus.updateInferenceSlaBreachMetrics.mockClear();
     await runEvaluate(service, 'conn-1', []);
-    // The Prometheus call must still include idx_a with breached=true so the
-    // time-series is continuous and Grafana does not see a resolve-like gap.
     expect(prometheus.updateInferenceSlaBreachMetrics).toHaveBeenCalledWith('conn-1', [
       { indexName: 'idx_a', breached: true },
     ]);
@@ -167,13 +139,9 @@ describe('InferenceLatencyService SLA state GC', () => {
     const { service } = await build({
       idx_a: { p99ThresholdUs: 5_000, enabled: true },
     });
-    // Establish a breach + debounce state.
     await runEvaluate(service, 'conn-1', [bucket({ p99: 9_000 })]);
     expect(internalState(service).has('conn-1|idx_a')).toBe(true);
 
-    // Next tick: no FT.SEARCH traffic (bucket disappears from the profile).
-    // State MUST NOT be GC'd — otherwise the next breach after traffic
-    // resumes would bypass the 10-min debounce and fire a duplicate webhook.
     await runEvaluate(service, 'conn-1', []);
     expect(internalState(service).has('conn-1|idx_a')).toBe(true);
   });
